backend/app/services/visual_search.py
"""
Visual Search Service using CLIP Model

This service enables users to upload images of cricket equipment
and find similar products in the catalog using visual similarity.

Note: This requires the transformers and torch libraries.
Install with: pip install transformers torch pillow
"""

from typing import List, Dict, Optional
import io
import structlog
from PIL import Image
import numpy as np
from fastapi import Depends

from app.db.database import Database, get_db

logger = structlog.get_logger()

# Flag to check if dependencies are available
try:
    from transformers import CLIPProcessor, CLIPModel
    import torch
    VISUAL_SEARCH_AVAILABLE = True
except ImportError:
    VISUAL_SEARCH_AVAILABLE = False
    logger.warning("visual_search_dependencies_missing", 
                   message="Install transformers, torch, and pillow for visual search")


class VisualSearchService:
    """Service for visual product search using CLIP model"""
    
    def __init__(self, db=None):
        self.db = db
        self.model = None
        self.processor = None
        self.device = None
        
        if VISUAL_SEARCH_AVAILABLE:
            self._initialize_model()
        else:
            logger.warning("visual_search_not_available", 
                          message="Visual search service initialized but dependencies missing")
    
    def _initialize_model(self):
        """Initialize CLIP model and processor"""
        try:
            logger.info("initializing_clip_model")
            
            # Use CPU by default (can be changed to CUDA if GPU available)
            self.device = "cuda" if torch.cuda.is_available() else "cpu"
            
            # Load CLIP model (using smaller variant for faster inference)
            model_name = "openai/clip-vit-base-patch32"
            self.model = CLIPModel.from_pretrained(model_name).to(self.device)
            self.processor = CLIPProcessor.from_pretrained(model_name)
            
            # Set model to evaluation mode
            self.model.eval()
            
            logger.info(
                "clip_model_initialized",
                model=model_name,
                device=self.device
            )
        except Exception as e:
            logger.error("clip_model_initialization_failed", error=str(e))
            self.model = None
            self.processor = None
    
    def is_available(self) -> bool:
        """Check if visual search is available"""
        return VISUAL_SEARCH_AVAILABLE and self.model is not None
    
    async def find_similar_products(
        self,
        image_bytes: bytes,
        limit: int = 5,
        category: Optional[str] = None
    ) -> List[Dict]:
        """
        Find products visually similar to the uploaded image.
        
        Args:
            image_bytes: Raw image bytes
            limit: Maximum number of results to return
            category: Optional category filter
        
        Returns:
            List of similar products with similarity scores
        """
        if not self.is_available():
            logger.error("visual_search_not_available")
            return []
        
        try:
            # Load and preprocess image
            image = Image.open(io.BytesIO(image_bytes)).convert("RGB")
            
            # Generate image embedding
            image_embedding = await self._generate_image_embedding(image)
            
            # Search for similar products in database
            similar_products = await self._search_similar_products(
                image_embedding,
                limit=limit,
                category=category
            )
            
            logger.info(
                "visual_search_completed",
                results_count=len(similar_products),
                category=category
            )
            
            return similar_products
            
        except Exception as e:
            logger.error("visual_search_failed", error=str(e))
            return []
    
    async def _generate_image_embedding(self, image: Image.Image) -> np.ndarray:
        """Generate CLIP embedding for an image"""
        try:
            # Preprocess image
            inputs = self.processor(images=image, return_tensors="pt").to(self.device)
            
            # Generate embedding
            with torch.no_grad():
                image_features = self.model.get_image_features(**inputs)
            
            # Normalize embedding
            image_features = image_features / image_features.norm(dim=-1, keepdim=True)
            
            # Convert to numpy array
            embedding = image_features.cpu().numpy()[0]
            
            return embedding
            
        except Exception as e:
            logger.error("generate_image_embedding_failed", error=str(e))
            raise
    
    async def _search_similar_products(
        self,
        query_embedding: np.ndarray,
        limit: int = 5,
        category: Optional[str] = None
    ) -> List[Dict]:
        """
        Search for products with similar embeddings.
        
        Note: This is a simplified implementation. In production, you would:
        1. Pre-compute embeddings for all product images
        2. Store embeddings in a vector database (e.g., pgvector, Pinecone, Weaviate)
        3. Use efficient similarity search (cosine similarity, FAISS)
        
        For now, this returns products based on category and rating.
        """
        if not self.db:
            logger.warning("database_not_available_for_visual_search")
            return []
        
        try:
            # TODO: Implement actual vector similarity search
            # For MVP, fall back to category-based search
            query = """
            SELECT id, name, price, rating, image_url, category, specifications
            FROM products
            WHERE in_stock = true
            """
            
            params = []
            if category:
                query += " AND category = $1"
                params.append(category)
            
            query += " ORDER BY rating DESC, sales_count DESC LIMIT $" + str(len(params) + 1)
            params.append(limit)
            
            results = await self.db.fetch_all(query, *params)
            
            products = []
            for row in results:
                products.append({
                    "product_id": row["id"],
                    "name": row["name"],
                    "price": float(row["price"]),
                    "rating": float(row["rating"]),
                    "image_url": row["image_url"],
                    "category": row["category"],
                    "similarity_score": 0.85,  # Placeholder score
                    "reason": "Visually similar cricket equipment"
                })
            
            return products
            
        except Exception as e:
            logger.error("search_similar_products_failed", error=str(e))
            return []
    
    async def generate_text_embedding(self, text: str) -> np.ndarray:
        """
        Generate CLIP embedding for text.
        Useful for text-based product search with visual similarity.
        """
        if not self.is_available():
            raise RuntimeError("Visual search not available")
        
        try:
            # Preprocess text
            inputs = self.processor(text=[text], return_tensors="pt", padding=True).to(self.device)
            
            # Generate embedding
            with torch.no_grad():
                text_features = self.model.get_text_features(**inputs)
            
            # Normalize embedding
            text_features = text_features / text_features.norm(dim=-1, keepdim=True)
            
            # Convert to numpy array
            embedding = text_features.cpu().numpy()[0]
            
            return embedding
            
        except Exception as e:
            logger.error("generate_text_embedding_failed", error=str(e))
            raise
    
    async def index_product_image(
        self,
        product_id: int,
        image_url: str
    ) -> bool:
        """
        Pre-compute and store embedding for a product image.
        
        This should be called when:
        1. New product is added
        2. Product image is updated
        3. Batch indexing all products
        
        Note: Requires implementation of vector storage (pgvector extension or external vector DB)
        """
        # TODO: Implement product image indexing
        # Steps:
        # 1. Download image from image_url
        # 2. Generate embedding
        # 3. Store embedding in vector database with product_id
        
        logger.info(
            "product_image_indexing_not_implemented",
            product_id=product_id,
            message="Implement vector storage for production use"
        )
        return False
    
    async def batch_index_products(self, product_ids: Optional[List[int]] = None):
        """
        Batch index multiple products for visual search.
        
        This should be run:
        1. During initial setup
        2. Periodically (e.g., nightly) to index new products
        """
        # TODO: Implement batch indexing
        logger.info("batch_indexing_not_implemented")
        pass


async def get_visual_search_service(db: Database = Depends(get_db)):
    """Dependency injection for VisualSearchService"""
    return VisualSearchService(db)
