backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
chromadb>=0.5.0
asyncpg>=0.29.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
slowapi>=0.1.9
python-multipart>=0.0.12
pydantic[email]>=2.10.0
structlog>=24.4.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
resend>=2.0.0
tenacity>=8.2.0
PyJWT>=2.8.0
sentry-sdk>=1.40.0

# Visual Search dependencies (optional)
transformers>=4.30.0
torch>=2.0.0
pillow>=10.0.0
